[Feat] #127 - 배너 뷰 연결
