+++ pophory-iOS/Screen/Mypage/View/MyPage/MyPageRootView.swift
@@ -46,12 +46,6 @@ class MyPageRootView: UIView {
     private lazy var adView: UIView = { UIView() }()
     private lazy var adEmptyView: UIImageView = { UIImageView(image: ImageLiterals.defaultBannerAd) }()
     
-    private lazy var feedTitleLabel: UILabel = { createFeedTitleLabel() }()
-    private lazy var emptyStackView: UIStackView = { createEmptyStackView() }()
-    private lazy var emptyImageView: UIImageView = { UIImageView(image: ImageLiterals.emptyFeedIcon) }()
-    private lazy var emptyDescriptionLabel: UILabel = { createEmptyDescriptionLabel() }()
-    private lazy var feedCollectionView: UICollectionView = { createFeedCollectionView() }()
-    
     // MARK: - Life Cycle
     
     override init(frame: CGRect) {
@@ -76,7 +70,6 @@ extension MyPageRootView {
         setupProfileView()
         setupBannerView()
         setupAdView()
-        setupFeedView()
     }
     
     private func setupHeaderView() {
@@ -178,46 +171,12 @@ extension MyPageRootView {
         
         adView.snp.makeConstraints { make in
             make.top.equalTo(storyBannerView.snp.bottom).offset(22)
-            make.leading.trailing.equalToSuperview().inset(28)
+            make.leading.trailing.bottom.equalToSuperview().inset(28)
             make.height.equalTo(100)
         }
         
         adEmptyView.snp.makeConstraints { make in
             make.edges.equalToSuperview()
-        }
-    }
-    
-    private func setupFeedView() {
-        contentView.addSubviews([
-            feedTitleLabel,
-            emptyStackView,
-            feedCollectionView
-        ])
-        
-        emptyStackView.addArrangedSubviews([
-            emptyImageView,
-            emptyDescriptionLabel
-        ])
-        
-        feedTitleLabel.snp.makeConstraints { make in
-            make.top.equalTo(adView.snp.bottom).offset(26)
-            make.leading.equalToSuperview().offset(20)
-        }
-        
-        emptyStackView.snp.makeConstraints { make in
-            make.top.equalTo(feedTitleLabel.snp.bottom).offset(46)
-            make.centerX.equalToSuperview()
-        }
-
-        emptyImageView.snp.makeConstraints { make in
-            make.size.equalTo(180)
-        }
-        
-        feedCollectionView.snp.makeConstraints { make in
-            make.top.equalTo(feedTitleLabel.snp.bottom).offset(12)
-            make.leading.trailing.equalTo(contentView).inset(20)
-            make.height.equalTo(0) // NOTE: don't erase this
-            make.bottom.equalTo(contentView).inset(350)
         }
     }
     
@@ -330,59 +289,6 @@ extension MyPageRootView {
         return view
     }
     
-    private func createFeedTitleLabel() -> UILabel {
-        let label = UILabel()
-        
-        label.font = .h2
-        label.text = "네컷사진 모아보기"
-        
-        return label
-    }
-    
-    private func createEmptyStackView() -> UIStackView {
-        let stackView = UIStackView()
-        
-        stackView.axis = .vertical
-        stackView.spacing = 16
-        stackView.alignment = .center
-        
-        return stackView
-    }
-    
-    private func createEmptyDescriptionLabel() -> UILabel {
-        let label = UILabel()
-        
-        label.font = .h3
-        label.textColor = .pophoryGray500
-        label.text = "네컷 사진을 추가해볼까?"
-        
-        return label
-    }
-    
-    private func createFeedCollectionView() -> UICollectionView {
-        let layout = UICollectionViewFlowLayout()
-        layout.minimumLineSpacing = 2
-        layout.minimumInteritemSpacing = 2
-        
-        layout.scrollDirection = .vertical
-        layout.sectionInset = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
-        
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.showsVerticalScrollIndicator = false
-        collectionView.showsHorizontalScrollIndicator = false
-        collectionView.isScrollEnabled = false
-        
-        collectionView.backgroundColor = .pophoryWhite
-        
-        collectionView.register(cell: PhotoCollectionViewCell.self)
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        
-        collectionView.showAnimatedGradientSkeleton()
-        
-        return collectionView
-    }
-    
     // MARK: - Logics
     
     @objc private func onClickSetting() {
@@ -390,11 +296,11 @@ extension MyPageRootView {
     }
     
     @objc private func onClickShare() {
-        delegate?.handleOnclickSetting()
+        delegate?.handleOnClickShare()
     }
     
     @objc private func onClickStory() {
-        delegate?.handleOnclickSetting()
+        delegate?.handleOnClickStory()
     }
     
     func updatePhotoCount(_ count: Int) {
@@ -411,63 +317,5 @@ extension MyPageRootView {
         }
         
         profileImageView.hideSkeleton()
-    }
-    
-    func updatePhotoData(_ photoData: [String]) {
-        self.photoData = photoData
-        
-        if photoData.isEmpty {
-            feedCollectionView.isHidden = true
-        } else {
-            feedCollectionView.isHidden = false
-            feedCollectionView.reloadData()
-            feedCollectionView.hideSkeleton()
-            
-            feedCollectionView.performBatchUpdates(nil, completion: { result in
-                self.feedCollectionView.snp.updateConstraints { make in
-                    make.height.equalTo(self.feedCollectionView.contentSize.height)
-                    make.bottom.equalTo(self.contentView).inset(20)
-                }
-            })
-        }
-    }
-}
-
-extension MyPageRootView: SkeletonCollectionViewDataSource {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return photoData?.count ?? 0
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: PhotoCollectionViewCell.identifier, for: indexPath) as? PhotoCollectionViewCell,
-              let photoData = photoData else {
-            return UICollectionViewCell()
-        }
-        
-        cell.configCell(imageUrl: photoData[indexPath.item], cellType: .myPage)
-        cell.clipsToBounds = true
-        cell.contentView.isSkeletonable = true
-        
-        return cell
-    }
-    
-    func collectionSkeletonView(_ skeletonView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return UICollectionView.automaticNumberOfSkeletonItems
-    }
-    
-    func collectionSkeletonView(_ skeletonView: UICollectionView, cellIdentifierForItemAt indexPath: IndexPath) -> SkeletonView.ReusableCellIdentifier {
-        return PhotoCollectionViewCell.identifier
-    }
-    
-    func collectionSkeletonView(_ skeletonView: UICollectionView, skeletonCellForItemAt indexPath: IndexPath) -> UICollectionViewCell? {
-        skeletonView.dequeueReusableCell(withReuseIdentifier: PhotoCollectionViewCell.identifier, for: indexPath)
-    }
-}
-
-extension MyPageRootView: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let size = (collectionView.frame.width - 4) / 3
-        return CGSize(width: size, height: size)
     }
 }
